cffi-c-ref.lisp: (c-let): nop.  a note about cffi compatibility
first the cffi typedef story: consider

(cffi:defcstruct (vect :size 8)
  (x :float :offset 0)
  (y :float :offset 4))

(cffi:defctype vect (:struct vect))

(defparameter $a (cffi:foreign-alloc '(:struct vect) :count 1))
(setf (cffi:foreign-slot-value $a 'vect 'x) 1.0)
(setf (cffi:foreign-slot-value $a 'vect 'y) 2.0)

(cffi::translate-from-foreign $a (cffi::parse-type '(:struct vect)))
;; => (Y 2.0 X 1.0)

(cffi::translate-from-foreign $a (cffi::parse-type 'vect))
;; => foreign-pointer. in vanilla cffi (as of 2023-03-11 commit
;; 48eaef6f)

Note this is quirk in vanilla cffi.  cffi is dereferencing the pointer
as an aggregate plist struct if the type is specified as a '(:struct
vect) and just returns the pointer if the type is specified as
'vect (which is a typedef to (:struct vect).

(cffi::parse-type '(:struct vect)) ; => #<VECT-TCLASS VECT>

with Vanilla cffi/cffi, parsing the typedef results in a
"FOREIGN-TYPEDEF".

(cffi::parse-type 'vect) ;=>#<CFFI::FOREIGN-TYPEDEF VECT>

with enometh/cffi, it results in an "ENHANCED-TYPEDEF"
(cffi::parse-type 'vect) ;=>#<CFFI::ENHANCED-TYPEDEF VECT>

Which results in consistent behaviour -- aggregate structs and
typedefs to aggregate structs are always dereferenced as plists. This
is necessary when working with nested structs and with libffi.
(There was a PR submitted to cffi/cffi in 2020 to fix this quirk which
was reverted but is nevertheless correct.)

with enometh/cffi
(cffi::translate-from-foreign $a (cffi::parse-type 'vect))
;; => (X 1.0 Y 2.0)

====

Now to address users of cffi-c-ref with vanilla cffi. They may have
written the equivalent of

(cffi-c-ref:c-let ((a vect :alloc t :clear t))
   (foo a))

where FOO is a function that expects a cffi:pointer to the struct.
This will not work with enometh/cffi. Instead the above code should be
rewritten as

(cffi-c-ref:c-let ((a (:struct vect) :alloc t :clear t))
   (foo (a &)))

to be consistent.  Similarly c-with, c-val.

This patch merely indicates what is needed for bug-compatibility with
vanilla cffi if the buggy code forms cannot be corrected.